+++ CHANGELOG.md
@@ -8,9 +8,9 @@ NOTE: This version bumps the Realm file format to version 11. It is not possible
 * Added RemoteMongoClient functionality to `Realm.User`
 
 ### Fixed
-* <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-js/issues/????), since v?.?.?)
 * Added missing `SyncConfiguration.error` field in the typescript definitions.
 * Fixed `SSL handshake failed: premature end of input` on Windows ([#2975](https://github.com/realm/realm-js/issues/2975, since v10.0.0-beta.1)
+* Missing `toJSON` TS declaration added for `Realm.Object` ([2903](https://github.com/realm/realm-js/issues/2903))
 
 ### Compatibility
 * MongoDB Realm Cloud.